Frontend: - rename components to Snake case
