+++ frontend/src/components/RoomOption.tsx
