+++ frontend/src/pages/index.tsx
@@ -6,10 +6,10 @@ import Image from "next/image";
 import styles from "@/styles/Home.module.css";
 import socket from "@/lib/socket";
 import { Container, Grid } from "@mui/material";
-import TopNav from "@/components/top_nav";
+import TopNav from "@/components/TopNav";
 import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
-import PlayButton from "@/components/play_button";
+import PlayButton from "@/components/PlayButton";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemAvatar from "@mui/material/ListItemAvatar";
@@ -18,19 +18,11 @@ import DialogTitle from "@mui/material/DialogTitle";
 import Dialog from "@mui/material/Dialog";
 import Avatar from "@mui/material/Avatar";
 import { blue } from "@mui/material/colors";
-import Options from "../components/room_options";
+import Options from "../components/RoomOption";
 
 const choices = ["Create room", "Join room", "Join random"];
 
 const User: NextPage = () => {
-  const [message, setMessage] = useState("");
-  useEffect(() => {
-    socket.connect();
-    return () => {
-      socket.disconnect();
-    };
-  }, []);
-  const onClink = () => {};
   const [open, setOpen] = useState(false);
   const [selectedValue, setSelectedValue] = useState(choices[1]);
